Backend/requirements.txt
aiohttp
aiosignal
asgiref
async-timeout
attrs
certifi
charset-normalizer
colorama
Django
django-cors-headers
frozenlist
gunicorn
idna
multidict
openai
packaging
psycopg2
requests
sqlparse
tqdm
typing_extensions
tzdata
urllib3
yarl
psycopg2-binary
dj-database-url
whitenoise[brotli]
djangorestframework
djangorestframework-simplejwt
google-generativeai
adal
aiohttp
aiosignal
annotated-types
anyio
asgiref
asttokens
async-timeout
attrs
azure-common
azure-core
azure-graphrbac
azure-identity
blinker
Brotli
cachelib
cachetools
certifi
cffi
charset-normalizer
click
colorama
cryptography
decorator
defusedxml
distro
dj-database-url
Django
django-allauth
django-cors-headers
djangorestframework
djangorestframework-simplejwt
executing
Flask
Flask-Session
frozenlist
google-ai-generativelanguage
google-api-core
google-auth
google-generativeai
googleapis-common-protos
grpcio
grpcio-status
gunicorn
h11
httpcore
httpx
identity
idna
imageio
imageio-ffmpeg
ipython
isodate
itsdangerous
jedi
Jinja2
MarkupSafe
matplotlib-inline
moviepy
msal
msal-extensions
msrest
msrestazure
multidict
numpy
oauthlib
openai
packaging
pandas
parso
pillow
portalocker
proglog
prompt-toolkit
proto-plus
protobuf
psycopg2
psycopg2-binary
pure-eval
pyasn1
pyasn1-modules
pycparser
pydantic
pydantic_core
Pygments
PyJWT
python-dateutil
python-dotenv
python3-openid
pytz
pywin32
requests
requests-oauthlib
rsa
setuptools
six
sniffio
sqlparse
stack-data
tqdm
traitlets
typing_extensions
tzdata
urllib3
wcwidth
Werkzeug
whitenoise
yarl
